Refactor modal, pagination, table, tag stories
