+++ stories/components/modal/modal.sample.js
@@ -1,0 +1,23 @@
+export default `
+  <div class="ch-overlay is-visible">
+    <div class="ch-overlay-content">
+      <div class="ch-modal is-open">
+        <header class="ch-modal-header">
+          <h4 class="ch-modal-header-content">Lorem ipsum dolor sit amet</h4>
+          <button type="button" class="ch-modal-header-actions ch-button ch-button-icon ch-button--ghost ch-paddingless" role="close">
+            <i class="ch-icon fal fa-times"></i>
+          </button>
+        </header>
+
+        <main class="ch-modal-content">
+          <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor, maiores.</p>
+        </main>
+
+        <footer class="ch-modal-footer">
+          <button type="button" class="ch-button ch-button--primary">Primary</button>
+          <button type="button" class="ch-button ch-button--ghost">Cancel</button>
+        </footer>
+      </div>
+    </div>
+  </div>
+`
