+++ stories/components/pagination/pagination.sample.js
@@ -1,0 +1,35 @@
+export default `
+  <nav class="ch-pagination">
+    <a href="#" class="ch-pagination-item ch-pagination-item--previous">
+      <i class="ch-icon"></i>
+    </a>
+
+    <a href="#" class="ch-pagination-item ch-pagination-item--first">
+      1
+    </a>
+
+    <button type="button" class="ch-pagination-item ch-pagination-item--gap">
+      <i class="ch-icon"></i>
+    </button>
+
+    <a href="#" class="ch-pagination-item" title="4">
+      4
+    </a>
+
+    <a href="#" class="ch-pagination-item" title="5">
+      5
+    </a>
+
+    <button type="button" class="ch-pagination-item ch-pagination-item--gap">
+      <i class="ch-icon"></i>
+    </button>
+
+    <a href="#" class="ch-pagination-item ch-pagination-item--last">
+      120
+    </a>
+
+    <a href="#" class="ch-pagination-item ch-pagination-item--next">
+      <i class="ch-icon"></i>
+    </a>
+  </nav>
+`
