+++ stories/components/tag/tags.sample.js
@@ -1,0 +1,3 @@
+export default `
+  <span class="ch-tag ch-tag--info">info</span>
+`
